requirements.txt
Django==1.3.1
Pillow==1.7.5
boto==2.1.1
django-storages==1.1.3
virtualenv==1.6.4
wsgiref==0.1.2
